requirements.txt
absl-py==0.12.0
alabaster==0.7.12
argon2-cffi==20.1.0
arrow==1.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bibtexparser==1.2.0
bleach==3.3.0
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
clang==5.0
click==8.0.0
cso-classifier==3.0
cryptography==3.4.7
cycler==0.10.0
cymem==2.0.5
Cython==0.29.23
decorator==5.0.7
defusedxml==0.7.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
docutils==0.16
entrypoints==0.3
fake-useragent==0.1.11
filelock==3.0.12
flatbuffers==1.12
forcelayout==1.0.6
free-proxy==1.0.2
future==0.18.2
gast==0.4.0
gensim==4.0.1
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
hdbscan==0.8.27
huggingface-hub==0.0.12
hurry.filesize==0.9
idna==2.10
imagesize==1.2.0
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
jupytext==1.11.2
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kneed==0.3.1
llvmlite==0.36.0
lxml==4.6.3
Markdown==3.3.4
markdown-it-py==1.1.0
MarkupSafe==1.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mdit-py-plugins==0.2.8
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.3.0
numba==0.53.1
numpy==1.21.2
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
preshed==3.0.5
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.7.4
Pygments==2.9.0
pynndescent==0.5.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-igraph==0.9.1
python-Levenshtein==0.12.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.4.4
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.19.0
rsa==4.7.2
sacremoses==0.0.45
scholarly==1.2.0
scikit-learn==0.24.2
scipy==1.6.3
selenium==3.141.0
semanticscholar==0.1.6
Send2Trash==1.5.0
sentence-transformers==2.0.0
sentencepiece==0.1.95
six==1.15.0
sklearn==0.0
smart-open==3.0.0
snowballstemmer==2.1.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.5
Sphinx==4.0.1
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
srsly==2.4.1
stem==1.8.0
strsimpy==0.2.0
tenacity==7.0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.5
testpath==0.4.4
texttable==1.6.4
thinc==8.0.8
threadpoolctl==2.1.0
tokenizers==0.10.3
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
transformers==4.8.0
typer==0.3.2
typing-extensions==3.7.4.3
umap-learn==0.5.1
update-checker==0.18.0
urllib3==1.26.4
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.12.1
